Moved CustomSchemesHttpResponseRedirect in views/util.py
